feat: complete graph
